+++ src/antsibull_docs/schemas/collection_config.py
@@ -5,22 +5,11 @@
 # SPDX-FileCopyrightText: 2023, Ansible Project
 """Schemas for collection config files."""
 
-import typing as t
-
 import pydantic as p
 
 from sphinx_antsibull_ext.schemas.ansible_output_data import (
-    Postprocessor,
-    PostprocessorNameRef,
+    NonRefPostprocessor,
 )
-
-
-def _is_not_name_ref(value: Postprocessor) -> Postprocessor:
-    if isinstance(value, PostprocessorNameRef):
-        raise ValueError(
-            "Cannot define name reference postprocessors in collection config"
-        )
-    return value
 
 
 class ChangelogConfig(p.BaseModel):
@@ -33,9 +22,7 @@ class AnsibleOutputConfig(p.BaseModel):
     global_env: dict[str, str] = {}
 
     # Named postprocessors
-    global_postprocessors: dict[
-        str, t.Annotated[Postprocessor, p.AfterValidator(_is_not_name_ref)]
-    ] = {}
+    global_postprocessors: dict[str, NonRefPostprocessor] = {}
 
     @p.field_validator("global_env", mode="before")
     @classmethod
