+++ src/sphinx_antsibull_ext/schemas/ansible_output_data.py
@@ -47,6 +47,7 @@ class PostprocessorNameRef(p.BaseModel):
 
 
 Postprocessor = t.Union[PostprocessorCLI, PostprocessorNameRef]
+NonRefPostprocessor = t.Union[PostprocessorCLI]
 
 
 class AnsibleOutputData(p.BaseModel):
